+++ src/main/java/at/hannibal2/skyhanni/config/features/InventoryConfig.java
@@ -131,13 +131,6 @@ public class InventoryConfig {
     public boolean jacobFarmingContestHighlightRewards = true;
 
     @Expose
-    @ConfigOption(name = "Duplicate Hider", desc = "Hides duplicate farming contests in the Jacob inventory.")
-    @ConfigEditorBoolean
-    @ConfigAccordionId(id = 3)
-    @FeatureToggle
-    public boolean jacobFarmingContestHideDuplicates = true;
-
-    @Expose
     @ConfigOption(name = "Contest Time", desc = "Adds the real time format to the Contest description.")
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 3)
@@ -383,6 +376,16 @@ public class InventoryConfig {
     public List<Integer> itemNumberAsStackSize = new ArrayList<>(Arrays.asList(3, 9, 11, 12));
 
     @Expose
+    @ConfigOption(
+            name = "Quick Craft Confirmation",
+            desc = "Require Ctrl+Click to craft items that aren't often quick crafted " +
+                    "(e.g. armor, weapons, accessories). Sack items can be crafted normally."
+    )
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean quickCraftingConfirmation = false;
+
+    @Expose
     @ConfigOption(name = "Sack Name", desc = "Show an abbreviation of the sack name.")
     @ConfigEditorBoolean
     @FeatureToggle
@@ -415,5 +418,10 @@ public class InventoryConfig {
     @FeatureToggle
     public boolean highlightAuctions = true;
 
+    @Expose
+    @ConfigOption(name = "Shift Click Equipment", desc = "Makes normal clicks to shift clicks in equipment inventory")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean shiftClickForEquipment = false;
 
 }