+++ src/main/java/at/hannibal2/skyhanni/config/features/MiscConfig.java
@@ -535,6 +535,172 @@ public class MiscConfig {
 
     }
 
+
+    @Expose
+    @ConfigOption(name = "Highlight Party Members", desc = "")
+    @Accordion
+    public HighlightPartyMembers highlightPartyMembers = new HighlightPartyMembers();
+
+    public static class HighlightPartyMembers {
+
+        @Expose
+        @ConfigOption(name = "Enabled", desc = "Marking party members with a bright outline to better find them in the world.")
+        @ConfigEditorBoolean
+        @FeatureToggle
+        public boolean enabled = false;
+
+        @Expose
+        @ConfigOption(
+                name = "Outline Color",
+                desc = "The color to outline party members in."
+        )
+        @ConfigEditorColour
+        public String outlineColor = "0:245:85:255:85";
+
+    }
+
+
+    @Expose
+    @ConfigOption(name = "Compact Tab List", desc = "")
+    @Accordion
+    public CompactTabListConfig compactTabList = new CompactTabListConfig();
+
+    public static class CompactTabListConfig {
+        @Expose
+        @ConfigOption(name = "Enabled", desc = "Compacts the tablist to make it look much nicer like SBA did. Also " +
+                "doesn't break god-pot detection and shortens some other lines.")
+        //made tablist one word here so both searches will pick it up
+        @ConfigEditorBoolean
+        @FeatureToggle
+        public boolean enabled = false;
+
+        @Expose
+        @ConfigOption(name = "Hide Hypixel Adverts", desc = "Hides text from advertising the Hypixel server or store in the tablist.")
+        @ConfigEditorBoolean
+        public boolean hideAdverts = false;
+
+        @Expose
+        @ConfigOption(name = "Advanced Player List", desc = "")
+        @Accordion
+        public AdvancedPlayerList advancedPlayerList = new AdvancedPlayerList();
+
+        public static class AdvancedPlayerList {
+
+            @Expose
+            @ConfigOption(name = "Player Sort", desc = "Change the sort order of player names in the tab list.")
+            @ConfigEditorDropdown(values = {"Rank (Default)", "SB Level", "Name (Abc)", "Ironman/Bingo", "Party/Friends/Guild", "Random"})
+            @ConfigAccordionId(id = 1)
+            public int playerSortOrder = 0;
+
+            @Expose
+            @ConfigOption(name = "Invert Sort", desc = "Flip the player list order on its head (also works with default rank).")
+            @ConfigEditorBoolean
+            public boolean reverseSort = false;
+
+            @Expose
+            @ConfigOption(name = "Hide Player Icons", desc = "Hide the icons/skins of player in the tab list.")
+            @ConfigEditorBoolean
+            public boolean hidePlayerIcons = false;
+
+            @Expose
+            @ConfigOption(name = "Hide Rank Color", desc = "Hide the player rank color.")
+            @ConfigEditorBoolean
+            public boolean hideRankColor = false;
+
+            @Expose
+            @ConfigOption(name = "Hide Emblems", desc = "Hide the emblems behind the player name.")
+            @ConfigEditorBoolean
+            public boolean hideEmblem = false;
+
+            @Expose
+            @ConfigOption(name = "Hide Level", desc = "Hide the SkyBlock level numbers.")
+            @ConfigEditorBoolean
+            public boolean hideLevel = false;
+
+            @Expose
+            @ConfigOption(name = "Hide Level Brackets", desc = "Hide the gray brackets in front of and behind the level numbers.")
+            @ConfigEditorBoolean
+            public boolean hideLevelBrackets = false;
+
+            @Expose
+            @ConfigOption(name = "Level Color As Name", desc = "Use the color of the SkyBlock level for the player color.")
+            @ConfigEditorBoolean
+            public boolean useLevelColorForName = false;
+
+            @Expose
+            @ConfigOption(name = "Bingo Rank Number", desc = "Show the number of the bingo rank next to the icon. Useful if you are not so familar with bingo.")
+            @ConfigEditorBoolean
+            public boolean showBingoRankNumber = false;
+
+            @Expose
+            @ConfigOption(name = "Hide Factions", desc = "Hide the icon of the Crimson Isle Faction in the tab list.")
+            @ConfigEditorBoolean
+            public boolean hideFactions = false;
+
+            @Expose
+            @ConfigOption(name = "Mark Special Persons", desc = "Show speical icons behind the name of guild members, party members, friends, and marked players.")
+            @ConfigEditorBoolean
+            public boolean markSpecialPersons = false;
+
+            @Expose
+            @ConfigOption(name = "Mark SkyHanni Devs", desc = "Adds a §c:O §7behind the tablist name of SkyHanni's creators.")
+            @ConfigEditorBoolean
+            public boolean markSkyHanniDevs = false;
+        }
+    }
+
+    @Expose
+    @ConfigOption(name = "Kick Duration", desc = "")
+    @Accordion
+    public KickDurationConfig kickDuration = new KickDurationConfig();
+
+    public static class KickDurationConfig {
+
+        @Expose
+        @ConfigOption(
+                name = "Enabled",
+                desc = "Show in the Hypixel lobby since when you were last kicked from SkyBlock (" +
+                        "useful if you get blocked because of '§cYou were kicked while joining that server!§7')."
+        )
+        @ConfigEditorBoolean
+        @FeatureToggle
+        public boolean enabled = true;
+
+        @Expose
+        @ConfigOption(name = "Warn Time", desc = "Send warning and sound this seconds after a SkyBlock kick.")
+        @ConfigEditorSlider(
+                minValue = 5,
+                maxValue = 300,
+                minStep = 1
+        )
+        public Property<Integer> warnTime = Property.of(60);
+
+        @Expose
+        public Position position = new Position(400, 200, 1.3f);
+    }
+
+    @Expose
+    @ConfigOption(name = "Player Chat Symbols", desc = "")
+    @Accordion
+    public ChatSymbols chatSymbols = new ChatSymbols();
+
+    public static class ChatSymbols {
+
+        @Expose
+        @ConfigOption(name = "Enabled", desc = "Adds extra symbols to the chat such as those from ironman, " +
+                "stranded, bingo or nether factions and places them next to your regular player emblems. " +
+                "Also allows emblems without sb level in chat.")
+        @ConfigEditorBoolean
+        @FeatureToggle
+        public boolean enabled = true;
+
+        @Expose
+        @ConfigOption(name = "Chat Symbol Location", desc = "Determines where the symbols should go in chat in relation to the " +
+                "player's name. Hidden will hide all emblems from the chat. §eRequires above setting to be on to hide the symbols.")
+        @ConfigEditorDropdown(values = {"Left", "Right", "Hidden"})
+        public int symbolLocation = 0;
+    }
+
     @Expose
     @ConfigOption(name = "Exp Bottles", desc = "Hides all the experience orbs lying on the ground.")
     @ConfigEditorBoolean
@@ -637,6 +803,17 @@ public class MiscConfig {
     @ConfigEditorBoolean
     @FeatureToggle
     public boolean fixNeuHeavyPearls = true;
+
+    @Expose
+    @ConfigOption(
+            name = "Time In Limbo",
+            desc = "Show the time since you entered the limbo.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean showTimeInLimbo = true;
+
+    @Expose
+    public Position showTimeInLimboPosition = new Position(400, 200, 1.3f);
 
     @Expose
     public Position inventoryLoadPos = new Position(394, 124, false, true);