+++ src/main/java/at/hannibal2/skyhanni/events/RepositoryReloadEvent.kt
@@ -1,7 +1,7 @@
 package at.hannibal2.skyhanni.events
 
 import at.hannibal2.skyhanni.data.repo.RepoUtils
-import at.hannibal2.skyhanni.test.command.CopyErrorCommand
+import at.hannibal2.skyhanni.test.command.ErrorManager
 import com.google.gson.Gson
 import com.google.gson.JsonObject
 import java.io.File
@@ -13,9 +13,9 @@ class RepositoryReloadEvent(val repoLocation: File, val gson: Gson) : LorenzEven
     inline fun <reified T : Any> getConstant(constant: String, type: Type? = null) = try {
         RepoUtils.getConstant(repoLocation, constant, gson, T::class.java, type)
     } catch (e: Exception) {
-        CopyErrorCommand.logError(
-                Exception("Repo parsing error while trying to read constant '$constant'", e),
-                "Error reading repo data"
+        ErrorManager.logError(
+            Exception("Repo parsing error while trying to read constant '$constant'", e),
+            "Error reading repo data"
         )
         null
     }