+++ src/main/java/at/hannibal2/skyhanni/features/fishing/FishingHookDisplay.kt
@@ -4,6 +4,7 @@ import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.events.CheckRenderEntityEvent
 import at.hannibal2.skyhanni.events.GuiRenderEvent
 import at.hannibal2.skyhanni.events.LorenzTickEvent
+import at.hannibal2.skyhanni.events.LorenzWorldChangeEvent
 import at.hannibal2.skyhanni.utils.EntityUtils
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.LorenzVec
@@ -11,38 +12,63 @@ import at.hannibal2.skyhanni.utils.RenderUtils.renderString
 import at.hannibal2.skyhanni.utils.getLorenzVec
 import net.minecraft.client.entity.EntityPlayerSP
 import net.minecraft.entity.item.EntityArmorStand
-import net.minecraft.entity.item.EntityXPOrb
 import net.minecraft.entity.projectile.EntityFishHook
 import net.minecraftforge.event.entity.EntityJoinWorldEvent
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 class FishingHookDisplay {
     private val config get() = SkyHanniMod.feature.fishing.fishingHookDisplay
+    private var bobber: EntityFishHook? = null
     private var bobberLocation: LorenzVec? = null
     private var armorStand: EntityArmorStand? = null
+    private val potentionArmorStands = mutableListOf<EntityArmorStand>()
     private var display = ""
+
+    @SubscribeEvent
+    fun onWorldChange(event: LorenzWorldChangeEvent) {
+        display = ""
+        bobberLocation = null
+        bobber = null
+        armorStand = null
+        potentionArmorStands.clear()
+    }
 
     @SubscribeEvent
     fun onTick(event: LorenzTickEvent) {
         if (!isEnabled()) return
-        if (!event.repeatSeconds(1)) return
+        if (!event.isMod(config.debugUpdateInterval)) return
+
+        if (armorStand == null) {
+            val filter = potentionArmorStands.filter { it.hasCustomName() && it.hasCorrectName() }
+            if (filter.size == 1) {
+                armorStand = filter[0]
+            }
+        }
 
         val entities = EntityUtils.getEntities<EntityFishHook>()
-        bobberLocation = entities.firstOrNull { it.angler is EntityPlayerSP }?.getLorenzVec()
+        val foundBobber = entities.firstOrNull { it.angler is EntityPlayerSP }
+        if (foundBobber != bobber) {
+            bobber = foundBobber
+            potentionArmorStands.clear()
+            armorStand = null
+        }
+
+        if (bobber != null) {
+            bobberLocation = bobber?.getLorenzVec()
+        }
     }
 
     @SubscribeEvent
     fun onJoinWorld(event: EntityJoinWorldEvent) {
         if (!isEnabled()) return
         val entity = event.entity ?: return
-        if (entity is EntityXPOrb) return
+        if (entity !is EntityArmorStand) return
         val bobberLocation = bobberLocation ?: return
 
         val distance = entity.getLorenzVec().distance(bobberLocation)
-        if (distance > 0.1) return
-        if (entity is EntityArmorStand) {
-            armorStand = entity
-        }
+        if (distance > config.debugMaxDistance) return
+
+        potentionArmorStands.add(entity)
     }
 
     @SubscribeEvent
@@ -56,7 +82,7 @@ class FishingHookDisplay {
     }
 
     @SubscribeEvent
-    fun onRenderOverlay(event: GuiRenderEvent.GameOverlayRenderEvent) {
+    fun onRenderOverlay(event: GuiRenderEvent.GuiOverlayRenderEvent) {
         if (!isEnabled()) return
 
         val armorStand = armorStand ?: return
@@ -64,6 +90,17 @@ class FishingHookDisplay {
         if (!armorStand.hasCustomName()) return
 
         config.position.renderString(armorStand.name, posLabel = "Fishing Hook Display")
+    }
+
+    private fun EntityArmorStand.hasCorrectName(): Boolean {
+        if (name == "§c§l!!!") {
+            return true
+        }
+        if (name.startsWith("§e§l")) {
+            return true
+        }
+
+        return false
     }
 
     fun isEnabled() = LorenzUtils.inSkyBlock && config.enabled && FishingAPI.hasFishingRodInHand()