+++ src/main/java/at/hannibal2/skyhanni/features/misc/ServerRestartTitle.kt
@@ -2,7 +2,6 @@ package at.hannibal2.skyhanni.features.misc
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.data.ScoreboardData
-import at.hannibal2.skyhanni.data.TitleUtils
 import at.hannibal2.skyhanni.events.LorenzTickEvent
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.StringUtils.matchMatcher
@@ -28,7 +27,7 @@ class ServerRestartTitle {
                 val totalSeconds = minutes * 60 + seconds
                 if (totalSeconds > 120 && totalSeconds % 30 != 0) return
                 val time = TimeUtils.formatDuration(totalSeconds.toLong() * 1000)
-                TitleUtils.sendTitle("§cServer Restart in §b$time", 2.seconds)
+                LorenzUtils.sendTitle("§cServer Restart in §b$time", 2.seconds)
             }
         }
     }