+++ src/main/java/at/hannibal2/skyhanni/features/misc/TimeFeatures.kt
@@ -28,7 +28,7 @@ class TimeFeatures {
     }
 
     @SubscribeEvent
-    fun onRenderOverlay(event: GuiRenderEvent.GameOverlayRenderEvent) {
+    fun onRenderOverlay(event: GuiRenderEvent.GuiOverlayRenderEvent) {
         if (!LorenzUtils.inSkyBlock) return
 
         if (config.realTime) {