+++ CHANGELOG.md
@@ -4,17 +4,88 @@
 
 ### New Features
 
-### Events
++ Added Quick Craft Confirmation. - Cad
+    + Require Ctrl+Click to craft items that aren't often quick crafted (e.g. armor, weapons, accessories).
+    + Sack items can be crafted normally.
++ Added **Compact Tab List**.
+    + Compacts the tablist to make it look much nicer (old SBA feature, but fewer bugs). - CalMWolfs
+    + Option to hide Hypixel advertisment banners. - CalMWolfs
+    + Added **Advanced Player List**. - hannibal2
+        + Customize the player list (inside the tab list) in various ways.
+        + Change the sort order of players: Default, SkyBlock Level, alphabetical name, Iron Man first/bingo level,
+          party/friends/guild
+        + Option to hide different parts of the player list: Player skins/icons, Hypixel rank color, Emblems, SkyBlock
+          level
++ Added Kick Duration. - hannibal2
+    + Show in the Hypixel lobby since when you were last kicked from SkyBlock.
+    + Useful if you get blocked because of 'You were kicked while joining that server!'.
+    + Send a warning and sound after a custom amount of seconds.
++ Added Time In Limbo. - hannibal2
+    + Show the time since you entered limbo.
+    + Show a chat message for how long you were in limbo once you leave it.
++ Added Highlight Party Members. - Cad
+    + Marking party members with a bright outline to better find them in the world.
++ Added Shift Click Equipment. - Thunderblade73
+    + This removes the need to shift-click to swap the equipment items, without the annoying "pick up animation".
++ Adds chat symbols such as iron man/bingo/nether faction like SBA had/has. - CalMWolfs
+    + Will not break with emblems.
+    + Optional if left or right side of name.
+    + Should not break with other mods.
+
+#### Events
 
 + Highlight Jerries during the Jerrypoclaypse. - Erymanthus
++ Show waypoints for Baskets of the Halloween Event in the main Hypixel lobby. - Erymanthus
+    + Thanks Tobbbb for the coordinates!
+    + Support for hiding basket waypoints once you have clicked on them. - hannibal2
+    + Option to show only the closest basket. - hannibal2
 
 ### Changes
 
++ Allowing clicks on the farming weight overlay to open the Farming Profile of you or the next person in the
+  leaderboard, or manual reloading the farming weight. - hannibal2
++ Added "Burning Desert" as a tarantula slayer area. - hannibal2
++ Slayer features now work only when the slayer type matches the current SkyBlock area. - hannibal2
++ Made Fatal Tempo same as Chimera in Estimated Item Value. - jani
++ Added debug options for fishing hook display. - hannibal2
+    + This should help find values that the fishing hook display works 100% with.
++ Changed the color for the tab list Special Persons Mark.
++ Mark SkyHanni Devs in the tab list special.
++ Added buttons to change the format of the price and the number in the sack display. - HiZe
++ Made Smoldering same as Fatal Tempo and Chimera in Estimated Item Value. - jani
++ Added an option to change where to get the items from in the composter overlay: from the bazaar or from sacks. - HiZe
++ Added mouse button support for key binds. - CalMWolfs
+
 ### Bug Fixes
+
++ The yaw/pitch display does no longer show scientific notations at small values. - hannibal2
++ Fixed slayer RNG Meter problem to detect the selected item. - hannibal2
++ Fixed capitalization errors. - J10a1n15
++ Fixed a bug that hides the contest inventory menu. - CalMWolfs
++ Fixed an error with the tab list. - hannibal2
++ Fixed fishing hook display data not properly resetting on a world change. - hannibal2
++ Fixed an error in Quick Craft Confirmation. - hannibal2
++ Fixed a crash with future NEU versions because of renamed code. - hannibal2
++ Fixed double rendering of item tooltips with chat triggers. - nea
++ Fixed Sacks Display integer limit error. - HiZe & hannibal2
++ Fixed the vitality attribute is wrongly labeled as "mending" in Estimated Item Value. - walker
++ Fixed lever clicks getting highlighted in the water room. - hannibal2
+    + This solution might not work for iron man in full parties.
++ Fixed slayer profit tracker showing Bazaar/AH price as NPC price. - hannibal2
++ Fixed Hyper reforge/End Stone Geode detection. - hannibal2
++ Better limbo leave detection. - hannibal2
++ Fixed rare crash when trying to read neu config. - hannibal2
++ Fixed rare case where the visitor description is empty. - hannibal2
 
 #### Config
 
 + Fixed two typos in the config description. - Absterge
++ Fixed small typos in config. - Absterge
+
+#### Removed Features
+
++ Removed **Duplicate Hider**.
+    + Hypixel now fixed the bug themselves and hides duplicate farming contests in the Jacob inventory.
 
 ## Version 0.20
 
@@ -512,7 +583,7 @@
     + Added a workaround for a crash when refusing a visitor - hannibal2
 + Added support for new counter drops to dicer rng counter - ReyMaratov
 + Fixed composter inventory numbers after Hypixel changes - hannibal2
-+ Rng dicer chat hider now works without enabling drop counter
++ RNG dicer chat hider now works without enabling drop counter
 + Server restart timer no longer shows all the time if over 2 minutes
 + Fixed crazy rare drops not counting properly - ReyMaratov
 + Fixed individual attribute prices in estimated item value - nea